+++ src/modules/edc-demo/services/contract-definition-builder.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {ContractDefinitionDto} from '../../edc-dmgmt-client';
+import {ContractDefinitionDto, policyDefinitionId} from '../../edc-dmgmt-client';
 import {ContractDefinitionEditorDialogFormValue} from '../components/contract-definition-editor-dialog/contract-definition-editor-dialog-form-model';
 import {AssetProperties} from './asset-properties';
 
@@ -18,8 +18,8 @@ export class ContractDefinitionBuilder {
   ): ContractDefinitionDto {
     return {
       id: formValue.id!.trim(),
-      accessPolicyId: formValue.accessPolicy?.id!,
-      contractPolicyId: formValue.contractPolicy?.id!,
+      accessPolicyId: policyDefinitionId(formValue.accessPolicy!),
+      contractPolicyId: policyDefinitionId(formValue.contractPolicy!),
       criteria: [
         {
           operandLeft: AssetProperties.id,