+++ src/modules/edc-dmgmt-client/model/policyDefinition.ts
@@ -18,3 +18,11 @@ export interface PolicyDefinition {
     id: string;
 }
 
+/**
+ * Temporary workaround around EDC Milestone 6/7 API incompatibilities.
+ * @param policyDefinition polcicy definition
+ * @return policy definition ID
+ */
+export function policyDefinitionId(policyDefinition: PolicyDefinition): string {
+  return policyDefinition.id || (policyDefinition as any).uid || 'no-known-id-field-on-policy-definition';
+}